Move raise_for_status into request.py
